Search filters working.
Themes migrated to material 3 as part of a theme update to allow for visibility of text in filters.

+++ lib/services/monster_storage.dart
@@ -1,13 +1,13 @@
 class Monster{
 
   String? name;
-  int? cr;
+  double? cr;
 
   setName (String? n) {
     name = n;
   }
 
-  setTable (int? c) {
+  setTable (double? c) {
     cr = c;
   }
 
@@ -15,7 +15,7 @@ class Monster{
     return name;
   }
 
-  int? getCr () {
+  double? getCr () {
     return cr;
   }
